src/requirements.txt
python-docx==0.8.10
pandas==0.24.2
seaborn==0.10.0
numpy=1.18.1
scipy=1.4.1
